refactor(services): split github and batch modules, tighten fail-open policy
- Extract `github.repository-content`, `github.pull-request`, and `github.progress-issue` modules composed by `GitHubService`
- Split translation batch work into branch lifecycle, PR lifecycle, and `TranslationFileProcessor` modules
- Retry transient PR validity errors in discovery before fail-open inclusion
- Fail closed on unexpected CLD errors after retries; keep empty, short, or unidentifiable content fail-open
- Extend tests for PR validity retry, CLD failure handling, and PR validity propagation
