+++ src/app/services/github/github.content.ts
@@ -1,754 +1,2 @@
-import type { components } from "@octokit/openapi-types";
-import type { RestEndpointMethodTypes } from "@octokit/rest";
-
-import type { CommentBuilderService } from "@/app/services/comment-builder/";
-import type {
-	PatchedRepositoryTreeItem,
-	ProcessedFileResult,
-	PullRequestIssueCommentSnapshot,
-	PullRequestReviewSnapshot,
-	PullRequestStatus,
-	RepositoryMarkdownBlob,
-	TranslationProgressFileRef,
-} from "@/app/services/github/types";
-
-import type { SharedGitHubDependencies } from "./types";
-
-import { withRetry } from "@/app/clients/";
-import {
-	hasReportableProgressComment,
-	selectProgressCommentPayload,
-} from "@/app/services/comment-builder/progress-comment.util";
-import { isSafeTranslatablePath, logger } from "@/app/utils/";
-import { ApplicationError, ErrorCode } from "@/shared/errors/";
-
-import {
-	decodeRepositoryFileContent,
-	fetchRepositoryContent,
-	fetchRepositoryDefaultBranch,
-	isRepositoryContentNotFound,
-	readRepositoryFileSha,
-} from "./github-api.util";
-
-/** Pull request options */
-export interface PullRequestOptions {
-	/** Source branch name */
-	branch: string;
-
-	/** Pull request title */
-	title: string;
-
-	/** Pull request description */
-	body: string;
-
-	/** Target branch for PR */
-	baseBranch?: string;
-}
-
-/** Options for committing translation changes */
-export interface CommitTranslationOptions {
-	/** Target branch reference */
-	branch: RestEndpointMethodTypes["git"]["getRef"]["response"]["data"];
-
-	/** File path being committed on the fork */
-	file: Pick<RepositoryMarkdownBlob, "path">;
-
-	/** Translated content */
-	content: string;
-
-	/** Commit message */
-	message: string;
-}
-
-/** Max attempts to poll GitHub for a computed `mergeable` value before giving up */
-const MERGEABLE_POLL_MAX_ATTEMPTS = 3;
-
-/** Delay between `mergeable` polling attempts in milliseconds */
-const MERGEABLE_POLL_DELAY_MS = 2_000;
-
-/** Max attempts to fetch a pull request file list before failing the workflow */
-const PR_FILES_FETCH_MAX_ATTEMPTS = 3;
-
-/**
- * Content and pull request operations module for GitHub API.
- *
- * Handles file content retrieval, translation commits, and PR management.
- */
-export class GitHubContent {
-	private readonly logger = logger.child({ component: GitHubContent.name });
-
-	/** Cached upstream default branch ref for repeated `getFile` calls in one run */
-	private upstreamDefaultBranchRef: string | undefined;
-
-	constructor(
-		private readonly deps: SharedGitHubDependencies,
-		private readonly commentBuilder: CommentBuilderService,
-	) {}
-
-	/**
-	 * Rejects repository paths that could escape the translatable `src/` tree.
-	 *
-	 * @param path Repository-relative file path
-	 * @param operation Caller operation name for error attribution
-	 */
-	private assertSafeTranslatablePath(path: string, operation: string): void {
-		if (isSafeTranslatablePath(path)) return;
-
-		throw new ApplicationError(
-			`Unsafe translatable path rejected: ${path}`,
-			ErrorCode.ResourceLoadError,
-			operation,
-			{ path },
-		);
-	}
-
-	/**
-	 * Creates a comment on a pull request.
-	 *
-	 * @param prNumber Pull request number
-	 * @param comment Comment to create
-	 *
-	 * @returns The response from the GitHub API
-	 */
-	public async createCommentOnPullRequest(
-		prNumber: number,
-		comment: string,
-	): Promise<RestEndpointMethodTypes["issues"]["createComment"]["response"]> {
-		const response = await this.deps.octokit.issues.createComment({
-			...this.deps.repositories.upstream,
-			issue_number: prNumber,
-			body: comment,
-		});
-
-		this.logger.info({ prNumber, commentId: response.data.id }, "Comment created on pull request");
-
-		return response;
-	}
-
-	/**
-	 * Lists all open pull requests. uses `octokit.paginate` to fetch all PRs.
-	 *
-	 * @returns A list of open pull requests
-	 */
-	public async listOpenPullRequests(): Promise<
-		RestEndpointMethodTypes["pulls"]["list"]["response"]["data"]
-	> {
-		const response = await this.deps.octokit.paginate("GET /repos/{owner}/{repo}/pulls", {
-			...this.deps.repositories.upstream,
-			state: "open",
-			per_page: 100,
-		});
-
-		this.logger.debug({ count: response.length }, "Listed open pull requests");
-
-		return response;
-	}
-
-	/**
-	 * Retrieves the list of files changed in a pull request. uses `octokit.paginate` to fetch all files.
-	 *
-	 * @param prNumber Pull request number to fetch changed files from
-	 *
-	 * @returns A `Promise` resolving to an array of file paths changed in the PR
-	 *
-	 * @example
-	 * ```typescript
-	 * const files = await content.getPullRequestFiles(123);
-	 * console.log(files);
-	 * // ^? ['src/content/learn/state-management.md', 'src/content/learn/hooks.md']
-	 * ```
-	 */
-	public async getPullRequestFiles(prNumber: number): Promise<string[]> {
-		const response = await withRetry(
-			() =>
-				this.deps.octokit.paginate("GET /repos/{owner}/{repo}/pulls/{pull_number}/files", {
-					...this.deps.repositories.upstream,
-					pull_number: prNumber,
-					per_page: 100,
-				}),
-			`${GitHubContent.name}.${this.getPullRequestFiles.name}`,
-			{ retries: PR_FILES_FETCH_MAX_ATTEMPTS - 1, minTimeout: 1, maxTimeout: 10, factor: 1 },
-		);
-
-		this.logger.debug({ count: response.length, prNumber }, "Retrieved pull request files");
-
-		return response.map((file) => file.filename);
-	}
-
-	/**
-	 * Commits translated content to a branch.
-	 *
-	 * Updates an existing file or creates a new one. Resolves the blob `sha` on the
-	 * target branch immediately before the commit so reused topic branches do not
-	 * send a stale tree `sha` (GitHub returns HTTP 409 when the `sha` does not match).
-	 *
-	 * @param options Commit options
-	 * @param options.branch Target branch ref for the commit
-	 * @param options.file File metadata including path and optional stale `sha`
-	 * @param options.content UTF-8 translated file body
-	 * @param options.message Commit message for the translation
-	 *
-	 * @returns Octokit `createOrUpdateFileContents` response
-	 *
-	 * @example
-	 * ```typescript
-	 * const options = {
-	 *   branch: branchRef,
-	 *   file: {
-	 *     path: 'src/content/homepage.md',
-	 *     content: translatedContent,
-	 *     sha: '1234567890',
-	 *     filename: 'homepage.md',
-	 *   },
-	 *   content: translatedContent,
-	 *   message: 'feat(i18n): translate homepage'
-	 * };
-	 *
-	 * await content.commitTranslation(options);
-	 * ```
-	 */
-	public async commitTranslation({
-		branch,
-		file,
-		content,
-		message,
-	}: CommitTranslationOptions): Promise<
-		RestEndpointMethodTypes["repos"]["createOrUpdateFileContents"]["response"]
-	> {
-		this.assertSafeTranslatablePath(
-			file.path,
-			`${GitHubContent.name}.${this.commitTranslation.name}`,
-		);
-
-		const blobShaOnBranch = await this.resolveBlobShaOnBranchForPath(branch.ref, file.path);
-
-		const response = await this.deps.octokit.repos.createOrUpdateFileContents({
-			...this.deps.repositories.fork,
-			path: file.path,
-			message,
-			content: Buffer.from(content).toString("base64"),
-			branch: branch.ref,
-			...(blobShaOnBranch !== undefined ? { sha: blobShaOnBranch } : {}),
-		});
-
-		this.logger.info(
-			{
-				filePath: file.path,
-				branch: branch.ref,
-				commitSha: response.data.commit.sha,
-			},
-			"Translation committed successfully",
-		);
-
-		return response;
-	}
-
-	/**
-	 * Looks up the current file blob `sha` on the fork at `branchRef`, or `undefined`
-	 * when the path is absent on that branch (create instead of update).
-	 *
-	 * @param branchRef Full ref such as `refs/heads/translate/foo`
-	 * @param path Repository path of the file
-	 *
-	 * @returns Blob `sha` on the branch, or `undefined` when the path is absent (create)
-	 */
-	private async resolveBlobShaOnBranchForPath(branchRef: string, path: string) {
-		try {
-			const existing = await fetchRepositoryContent(
-				this.deps.octokit,
-				this.deps.repositories.fork,
-				path,
-				branchRef,
-			);
-
-			if (Array.isArray(existing.data)) {
-				this.logger.warn(
-					{ path, branchRef },
-					"GitHub returned a directory listing for getContent; omitting sha for file update",
-				);
-
-				return undefined;
-			}
-
-			return readRepositoryFileSha(existing.data);
-		} catch (error) {
-			if (isRepositoryContentNotFound(error)) {
-				return undefined;
-			}
-
-			throw error;
-		}
-	}
-
-	/**
-	 * Creates a pull request.
-	 *
-	 * @param options Pull request options
-	 * @param options.branch Head branch name for the translation PR
-	 * @param options.title PR title shown on GitHub
-	 * @param options.body PR description body (markdown)
-	 * @param options.baseBranch Base branch on the upstream repo (default branch when omitted)
-	 *
-	 * @returns Octokit `pulls.create` response
-	 *
-	 * @example
-	 * ```typescript
-	 * const options = {
-	 *   branch: 'translate/homepage',
-	 *   title: 'feat(i18n): translate homepage',
-	 *   body: 'Translates homepage content to Portuguese',
-	 *   baseBranch: 'main',
-	 * };
-	 *
-	 * const pr = await content.createPullRequest(options);
-	 * ```
-	 */
-	public async createPullRequest({
-		branch,
-		title,
-		body,
-		baseBranch = "main",
-	}: PullRequestOptions): Promise<RestEndpointMethodTypes["pulls"]["create"]["response"]["data"]> {
-		const targetRepo = this.deps.repositories.upstream;
-		const headRef = `${this.deps.repositories.fork.owner}:${branch}`;
-
-		const createPullRequestResponse = await this.deps.octokit.pulls.create({
-			...targetRepo,
-			title,
-			body,
-			head: headRef,
-			base: baseBranch,
-			maintainer_can_modify: true,
-		});
-
-		this.logger.info(
-			{
-				prNumber: createPullRequestResponse.data.number,
-				title,
-				targetRepo: targetRepo.owner,
-				headRef,
-				baseBranch,
-			},
-			"Pull request created successfully",
-		);
-
-		return createPullRequestResponse.data;
-	}
-
-	/**
-	 * Updates the body of an open translation pull request on the upstream repository.
-	 *
-	 * @param prNumber Pull request number
-	 * @param body Markdown body (for example advisory validation warnings after re-translation)
-	 *
-	 * @returns Updated pull request data
-	 */
-	public async updatePullRequestBody(
-		prNumber: number,
-		body: string,
-	): Promise<RestEndpointMethodTypes["pulls"]["update"]["response"]["data"]> {
-		const response = await this.deps.octokit.pulls.update({
-			...this.deps.repositories.upstream,
-			pull_number: prNumber,
-			body,
-		});
-
-		this.logger.info({ prNumber }, "Pull request body updated");
-
-		return response.data;
-	}
-
-	/**
-	 * Fetches source markdown from the upstream default branch at `file.path`.
-	 *
-	 * Uses `repos.getContent` on the upstream repository so discovery always reads
-	 * source from `main` (or the upstream default), not bytes from a fork
-	 * `translate/...` branch that may already contain a translation.
-	 *
-	 * @param file File reference from the upstream repository tree
-	 *
-	 * @returns Upstream markdown blob content and `sha`
-	 */
-	public async getFile(file: PatchedRepositoryTreeItem): Promise<RepositoryMarkdownBlob> {
-		this.assertSafeTranslatablePath(file.path, `${GitHubContent.name}.${this.getFile.name}`);
-
-		const ref = await this.resolveUpstreamDefaultBranchRef();
-
-		const response = await fetchRepositoryContent(
-			this.deps.octokit,
-			this.deps.repositories.upstream,
-			file.path,
-			ref,
-		);
-
-		if (Array.isArray(response.data)) {
-			throw new ApplicationError(
-				`Expected file at path but received directory listing: ${file.path}`,
-				ErrorCode.ResourceLoadError,
-				`${GitHubContent.name}.${this.getFile.name}`,
-				{ path: file.path, ref },
-			);
-		}
-
-		const content = decodeRepositoryFileContent(response.data);
-
-		if (!content) {
-			throw new ApplicationError(
-				`Upstream file has no content: ${file.path}`,
-				ErrorCode.ResourceLoadError,
-				`${GitHubContent.name}.${this.getFile.name}`,
-				{ path: file.path, ref },
-			);
-		}
-
-		return {
-			content,
-			filename: file.filename,
-			path: file.path,
-			sha: response.data.sha || file.sha,
-		};
-	}
-
-	/**
-	 * Reads file content from the fork at a translation branch tip.
-	 *
-	 * @param path Repository path of the markdown file
-	 * @param branchName Translation branch name without `refs/heads/` prefix
-	 *
-	 * @returns File body as UTF-8 text, or `undefined` when the path is missing on that branch
-	 */
-	public async getForkFileContentAtBranch(path: string, branchName: string) {
-		const branchRef = `refs/heads/${branchName}`;
-
-		try {
-			const response = await fetchRepositoryContent(
-				this.deps.octokit,
-				this.deps.repositories.fork,
-				path,
-				branchRef,
-			);
-
-			if (Array.isArray(response.data)) {
-				this.logger.warn(
-					{ path, branchRef },
-					"GitHub returned a directory listing for fork branch content lookup",
-				);
-
-				return undefined;
-			}
-
-			return decodeRepositoryFileContent(response.data);
-		} catch (error) {
-			if (isRepositoryContentNotFound(error)) {
-				return undefined;
-			}
-
-			throw error;
-		}
-	}
-
-	/**
-	 * Resolves and caches the upstream repository default branch name.
-	 *
-	 * @returns Default branch name (for example `main`)
-	 */
-	private async resolveUpstreamDefaultBranchRef() {
-		if (this.upstreamDefaultBranchRef) {
-			return this.upstreamDefaultBranchRef;
-		}
-
-		const response = await fetchRepositoryDefaultBranch(
-			this.deps.octokit,
-			this.deps.repositories.upstream,
-		);
-		this.upstreamDefaultBranchRef = response;
-
-		return this.upstreamDefaultBranchRef;
-	}
-
-	/**
-	 * Finds the translation progress issue in the upstream repository.
-	 *
-	 * @returns The translation progress issue data or `undefined` if not found
-	 */
-	private async findTranslationProgressIssue(): Promise<
-		components["schemas"]["issue-search-result-item"] | undefined
-	> {
-		const queryString = `repo:${this.deps.repositories.upstream.owner}/${this.deps.repositories.upstream.repo} in:title "Translation Progress" is:issue is:open`;
-
-		const issueExistsResponse = await this.deps.octokit.rest.search.issuesAndPullRequests({
-			q: queryString,
-		});
-
-		if (issueExistsResponse.data.items.length > 1) {
-			this.logger.warn(
-				{ count: issueExistsResponse.data.items.length },
-				"Multiple translation progress issues found",
-			);
-
-			const correctIssue = issueExistsResponse.data.items.find((issue) => {
-				const possibleAssociations: components["schemas"]["author-association"][] = [
-					"OWNER",
-					"CONTRIBUTOR",
-					"COLLABORATOR",
-					"MEMBER",
-				];
-
-				return possibleAssociations.includes(issue.author_association);
-			});
-
-			if (!correctIssue) {
-				this.logger.error(
-					"Could not determine the correct translation progress issue from multiple candidates",
-				);
-
-				return undefined;
-			}
-
-			return correctIssue;
-		}
-
-		const translationProgressIssue = issueExistsResponse.data.items[0];
-
-		if (!translationProgressIssue) {
-			this.logger.warn("No translation progress issue found");
-
-			return undefined;
-		}
-
-		return translationProgressIssue;
-	}
-
-	/**
-	 * Posts translation results as comments on GitHub issues.
-	 *
-	 * ### Workflow
-	 *
-	 * 1. Returns early when there are no results, no candidate files, or no pull
-	 *    requests were opened or updated (avoids posting on failure-only runs)
-	 * 2. Resolves the translation progress issue on the upstream repository
-	 * 3. Creates a new issue comment with the compiled summary
-	 *
-	 * @param results Translation results to report
-	 * @param filesToTranslate Files that were translated
-	 *
-	 * @returns The comment created on the issue, or `undefined` when skipped or no issue
-	 *
-	 * @example
-	 * ```typescript
-	 * const comment = await content.commentCompiledResultsOnIssue(results, filesToTranslate);
-	 * ```
-	 */
-	public async commentCompiledResultsOnIssue(
-		results: ProcessedFileResult[],
-		filesToTranslate: readonly TranslationProgressFileRef[],
-	): Promise<RestEndpointMethodTypes["issues"]["createComment"]["response"]["data"] | undefined> {
-		if (results.length === 0 || filesToTranslate.length === 0) {
-			this.logger.warn("No results or files to translate. Skipping issue comment update");
-			return;
-		}
-
-		const payload = selectProgressCommentPayload(results, filesToTranslate);
-
-		if (!hasReportableProgressComment(payload)) {
-			this.logger.info(
-				{ resultCount: results.length },
-				"No pull requests were opened or updated in this run; skipping translation progress issue comment",
-			);
-
-			return;
-		}
-
-		const translationProgressIssue = await this.findTranslationProgressIssue();
-
-		if (!translationProgressIssue) {
-			this.logger.warn("Translation progress issue not found");
-			return;
-		}
-
-		const createCommentResponse = await this.deps.octokit.issues.createComment({
-			...this.deps.repositories.upstream,
-			issue_number: translationProgressIssue.number,
-			body: this.commentBuilder.buildProgressComment(payload),
-		});
-
-		this.logger.info(
-			{
-				issueNumber: translationProgressIssue.number,
-				commentId: createCommentResponse.data.id,
-			},
-			"Created comment on issue with compiled results",
-		);
-
-		return createCommentResponse.data;
-	}
-
-	/**
-	 * Retrieves a pull request by branch name.
-	 *
-	 * @param branchName Source branch name
-	 *
-	 * @returns The first matching pull request or undefined if none found
-	 */
-	public async findPullRequestByBranch(
-		branchName: string,
-	): Promise<RestEndpointMethodTypes["pulls"]["list"]["response"]["data"][number] | undefined> {
-		const response = await this.deps.octokit.pulls.list({
-			...this.deps.repositories.upstream,
-			head: `${this.deps.repositories.fork.owner}:${branchName}`,
-		});
-
-		return response.data[0];
-	}
-
-	/**
-	 * Checks if a pull request has merge conflicts that require closing and recreating.
-	 *
-	 * Polls GitHub when `mergeable` is `null` (async computation pending) to avoid
-	 * false negatives. After exhausting retries, treats an undetermined state as
-	 * conflicted to avoid silently skipping stale PRs.
-	 *
-	 * @param prNumber Pull request number to check
-	 *
-	 * @returns PR status with conflict and mergeability information
-	 */
-	public async checkPullRequestStatus(prNumber: number): Promise<PullRequestStatus> {
-		let pr = await this.fetchPullRequest(prNumber);
-
-		for (
-			let attempt = 0;
-			pr.mergeable === null && attempt < MERGEABLE_POLL_MAX_ATTEMPTS;
-			attempt++
-		) {
-			this.logger.debug(
-				{ prNumber, attempt: attempt + 1, maxAttempts: MERGEABLE_POLL_MAX_ATTEMPTS },
-				"PR mergeable state not yet computed, polling",
-			);
-			await new Promise((resolve) => setTimeout(resolve, MERGEABLE_POLL_DELAY_MS));
-			pr = await this.fetchPullRequest(prNumber);
-		}
-
-		if (pr.mergeable === null) {
-			this.logger.warn(
-				{ prNumber, mergeableState: pr.mergeable_state },
-				"PR mergeable state still undetermined after polling, treating as conflicted",
-			);
-		}
-
-		const hasConflicts =
-			pr.mergeable === null ||
-			(!pr.mergeable && (pr.mergeable_state === "dirty" || pr.mergeable_state === "unknown"));
-		const needsUpdate = hasConflicts;
-
-		return {
-			hasConflicts,
-			mergeable: pr.mergeable,
-			mergeableState: pr.mergeable_state,
-			needsUpdate,
-			createdBy: pr.user.login,
-		};
-	}
-
-	/**
-	 * Fetches a single PR's data from GitHub.
-	 *
-	 * @param prNumber Pull request number on the upstream repository
-	 *
-	 * @returns Pull request payload from the GitHub REST API
-	 */
-	private async fetchPullRequest(prNumber: number) {
-		const response = await this.deps.octokit.pulls.get({
-			...this.deps.repositories.upstream,
-			pull_number: prNumber,
-		});
-
-		return response.data;
-	}
-
-	/**
-	 * Lists issue comments on an open translation pull request (upstream repo).
-	 *
-	 * @param prNumber Pull request number on the upstream repository
-	 *
-	 * @returns Normalized issue comments, oldest first
-	 */
-	public async listPullRequestIssueComments(
-		prNumber: number,
-	): Promise<PullRequestIssueCommentSnapshot[]> {
-		const comments = await this.deps.octokit.paginate(
-			"GET /repos/{owner}/{repo}/issues/{issue_number}/comments",
-			{
-				...this.deps.repositories.upstream,
-				issue_number: prNumber,
-				per_page: 100,
-			},
-		);
-
-		return comments.map((comment) => ({
-			login: comment.user?.login ?? "",
-			authorAssociation: comment.author_association,
-			userType: comment.user?.type ?? "User",
-			createdAt: new Date(comment.created_at),
-			body: comment.body ?? "",
-		}));
-	}
-
-	/**
-	 * Lists submitted pull request reviews on an open translation pull request (upstream repo).
-	 *
-	 * @param prNumber Pull request number on the upstream repository
-	 *
-	 * @returns Normalized reviews, oldest first
-	 */
-	public async listPullRequestReviews(prNumber: number): Promise<PullRequestReviewSnapshot[]> {
-		const reviews = await this.deps.octokit.paginate(
-			"GET /repos/{owner}/{repo}/pulls/{pull_number}/reviews",
-			{
-				...this.deps.repositories.upstream,
-				pull_number: prNumber,
-				per_page: 100,
-			},
-		);
-
-		return reviews.flatMap((review) => {
-			if (!review.submitted_at) {
-				return [];
-			}
-
-			return [
-				{
-					id: review.id,
-					login: review.user?.login ?? "",
-					authorAssociation: review.author_association,
-					userType: review.user?.type ?? "User",
-					state: review.state as PullRequestReviewSnapshot["state"],
-					submittedAt: new Date(review.submitted_at),
-					body: review.body,
-				},
-			];
-		});
-	}
-
-	/**
-	 * Closes a pull request by number.
-	 *
-	 * @param prNumber Pull request number
-	 *
-	 * @returns Updated pull request data after setting `state` to `closed`
-	 *
-	 * @throws {Error} If pull request closure fails
-	 */
-	public async closePullRequest(
-		prNumber: number,
-	): Promise<RestEndpointMethodTypes["pulls"]["update"]["response"]["data"]> {
-		const response = await this.deps.octokit.pulls.update({
-			...this.deps.repositories.upstream,
-			pull_number: prNumber,
-			state: "closed",
-		});
-
-		this.logger.info({ prNumber }, "Pull request closed successfully");
-
-		return response.data;
-	}
-}
+export type { CommitTranslationOptions } from "./github.repository-content";
+export type { PullRequestOptions } from "./github.pull-request";
