+++ tests/services/runner/file-discovery.manager.spec.ts
@@ -43,12 +43,28 @@ describe("FileDiscoveryManager", () => {
 				[FAIL_OPEN_REASONS.prValidityEvaluationError]: 0,
 				[FAIL_OPEN_REASONS.languageDetectionEmptyContent]: 0,
 				[FAIL_OPEN_REASONS.languageDetectionShortContent]: 0,
-				[FAIL_OPEN_REASONS.languageDetectionCldError]: 0,
+				[FAIL_OPEN_REASONS.languageDetectionCldUnreliable]: 0,
 			};
 
 			await manager.filterByPRs([candidate], failOpenInventory);
 
 			expect(failOpenInventory[FAIL_OPEN_REASONS.prValidityEvaluationError]).toBe(1);
+		});
+
+		test("retries transient PR validity errors before fail-open", async () => {
+			const github = createMockGitHubService();
+			github.findPullRequestByBranch
+				.mockRejectedValueOnce(new Error("ECONNRESET"))
+				.mockRejectedValueOnce(new Error("ECONNRESET"))
+				.mockRejectedValue(new Error("GitHub API unavailable"));
+
+			const manager = createTestFileDiscoveryManager({ github });
+			const candidate = createRepositoryTreeItemFixture({ path: "src/content/page.md" });
+
+			const result = await manager.filterByPRs([candidate]);
+
+			expect(github.findPullRequestByBranch).toHaveBeenCalledTimes(3);
+			expect(result.filesToFetch).toHaveLength(1);
 		});
 
 		test("skips file when open translation pull request is valid", async () => {
